Got rid of the duplicated code
